vertex-ai-chatbot/requirements.txt
flask==3.0.0
google-cloud-aiplatform==1.38.0
google-cloud-storage==2.10.0
vertexai==0.0.1
chromadb==0.4.18
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
